aep-ot: model the sprite-command float slots faithfully, @complete drawSprite
The AepOtSpriteCmd pool slot is reinterpreted per command type, and the two
sprite variants encode the same offsets with OPPOSITE types. Verified from the
flush's per-case loads (renderAepOrderingTable FUN_000115d0): the type-0 sprite
(case 0) reads +0x1c/+0x20 as float (vldr 0x11622/0x11626) and +0x28/+0x2c as int
(ldr); the type-1 stretched sprite (case 1) reads +0x1c/+0x20 as int (ldr
0x11694/0x11698) and +0x28/+0x2c as float (vldr 0x116a0/0x116a4); cases 2-5 read
all four as int. drawSprite (FUN_00011468) stores +0x28/+0x2c with the only vstr
pair in the function and everything else with str.

A single field type cannot be bit-faithful to both sprite variants, so +0x1c,
+0x20, +0x28, and +0x2c are now unions of an int view (flPosXf/flPosYf/nOfsY/
nColorA) and a float view (flPosXfF/flPosYfF/nOfsYF/nColorAF). Each command writes
and reads the SAME member (no inactive-member punning), so every path matches the
binary's exact str/vstr:
- drawSprite (case 1) writes the float scales to nOfsYF/nColorAF and positions to
  the int flPosXf/flPosYf; its wrapper neTextureForiOS::draw now vcvt-casts exactly
  the two scale args (p.sx/p.sy) to float and passes the positions as ints,
  mirroring FUN_0000fbcc's conversions at 0x0fbf4/0x0fc0c.
- AepFrameDraw (case 0) writes the float scales to flPosXfF/flPosYfF and the size
  to the int nOfsY/nColorA.
- The flush reads each case through the view the binary loads.

Previously drawSprite stored ints where the binary stores floats (numerically
transparent via C++ value conversion, but not bit-faithful). drawSprite is now
marked @complete.

+++ Project/System/src/Aep/AepManager.mm
@@ -430,12 +430,12 @@ void AepManager::drawTransitionOverlay(int alpha) {
     }
     cmd->wFlags = 4; // -> renderAepOrderingTable case 4 = drawAepOtRect
     cmd->nBank = 0;
-    cmd->nTexU = m_transitionOverlay[0];     // x
-    cmd->nTexV = m_transitionOverlay[1];     // y
-    cmd->nPosX = m_transitionOverlay[2];     // w
-    cmd->nPosY = m_transitionOverlay[3];     // h
-    cmd->flPosXf = (float)alpha;             // drawAepOtRect nAlpha (0..100)
-    cmd->flPosYf = (float)m_transitionColor; // drawAepOtRect nColor (0x00RRGGBB; 0 = black)
+    cmd->nTexU = m_transitionOverlay[0]; // x
+    cmd->nTexV = m_transitionOverlay[1]; // y
+    cmd->nPosX = m_transitionOverlay[2]; // w
+    cmd->nPosY = m_transitionOverlay[3]; // h
+    cmd->flPosXf = alpha;                // drawAepOtRect nAlpha (0..100); +0x1c int slot
+    cmd->flPosYf = m_transitionColor;    // drawAepOtRect nColor (0x00RRGGBB; 0 = black); +0x20 int
 }
 
 // ===========================================================================
